Moved login logic - also moved sections selector to ethereum/redux/selectors
