+++ src/redux/session/thunks.js
@@ -3,12 +3,21 @@
  * Licensed under the AGPL Version 3 license.
  */
 
+import { getAddress } from '@chronobank/bitcoin/utils'
+import { bitcoinCreateWallet } from '@chronobank/bitcoin/redux/actions'
+import { ethereumCreateWallet } from '@chronobank/ethereum/redux/actions'
+import { getPrivateKeyByMnemonic, getAddressByMnemonic } from '@chronobank/ethereum/utils'
 import { login, logout, savePrivateKey } from './actions'
 
-export const loginThunk = (privateKey, currentWallet) => (dispatch) => {
+export const loginThunk = (mnemonic) => (dispatch) => {
   try {
+    const privateKey = getPrivateKeyByMnemonic(mnemonic)
+    const bitcoinAddress = getAddress(privateKey)
+    const ethAddress = getAddressByMnemonic(mnemonic)
+    dispatch(bitcoinCreateWallet(bitcoinAddress))
+    dispatch(ethereumCreateWallet(ethAddress))
     dispatch(savePrivateKey(privateKey))
-    dispatch(login(currentWallet))
+    dispatch(login(ethAddress))
   } catch (e) {
     return Promise.reject(e)
   }
